+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * Changed: Translation on Datepickers now is german as well
 * Fixed: When user is teacher, needed users are not requested
+- Fixed: Translation of Datatable
 
 ## [v0.8.7] - 2018-11-21